+++ Runtime/ObjectHandling/TransformData.cs
@@ -50,7 +50,7 @@ namespace Spellbound.Core {
                     Scale = Scale
                 };
 
-        public Quaternion RotAsQuaternion() => Quaternion.Euler(Rotation);
+        public Quaternion RotAsQuaternion() => quaternion.EulerXYZ(math.radians(Rotation));
 
         public Vector3 ScaleAsVector3() => new(Scale, Scale, Scale);